+++ src/gemseo_umdo/formulations/_statistics/control_variate/variance.py
@@ -14,14 +14,16 @@
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """Estimators of the variance for U-MDO formulation based on control variates."""
 
-from gemseo.typing import RealArray
-from numpy import diag
-from numpy import diagonal
-from numpy.linalg import multi_dot
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
 
 from gemseo_umdo.formulations._statistics.control_variate.base_control_variate_estimator import (  # noqa: E501
     BaseControlVariateEstimator,
 )
+
+if TYPE_CHECKING:
+    from gemseo.typing import RealArray
 
 
 class Variance(BaseControlVariateEstimator):
@@ -29,14 +31,13 @@ class Variance(BaseControlVariateEstimator):
 
     def estimate_statistic(  # noqa: D102
         self,
-        samples: RealArray,
-        u_samples: RealArray,
+        evaluations: RealArray,
         mean: RealArray,
-        jac: RealArray,
+        variance: RealArray,
+        some_predictions: RealArray,
+        many_predictions: RealArray,
     ) -> RealArray:
-        cv_samples = self._compute_control_variate_samples(u_samples, mean, jac)
-        diff2 = (samples - samples.mean()) ** 2
-        cv_diff2 = (cv_samples - cv_samples.mean()) ** 2
-        cv_var = diagonal(multi_dot([jac, diag(self._u_standard_deviation**2), jac.T]))
+        diff2 = (evaluations - evaluations.mean(0)) ** 2
+        cv_diff2 = (some_predictions - some_predictions.mean(0)) ** 2
         alpha = self._compute_opposite_scaled_covariance(diff2, cv_diff2)
-        return (diff2.mean(0) + alpha * (cv_diff2.mean(0) - cv_var)).ravel()
+        return (diff2.mean(0) + alpha * (cv_diff2.mean(0) - variance)).ravel()