+++ CHANGELOG.md
@@ -26,6 +26,31 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 4.2.0 (August 2025)
+
+### Added
+
+- The [ControlVariate][gemseo_umdo.formulations.control_variate.ControlVariate] U-MDO formulation can create the control variates from any [BaseRegressor][gemseo.mlearning.regression.algos.base_regressor.BaseRegressor] using the arguments `regressor_settings` (for defining the regressor) and `regressor_doe_algo_settings` (for creating the training dataset). By default, this U-MDO formulation uses control variates based on Taylor polynomials.
+- The subpackage [truss][gemseo_umdo.use_cases.truss] includes modules for implementing a truss structure problem from the literature.
+- The function [create_noising_discipline_chain][gemseo_umdo.disciplines.utils.create_noising_discipline_chain] returns a disciplines chain to noise input variables.
+  This function is used by the [UDOEScenario][gemseo_umdo.scenarios.udoe_scenario.UDOEScenario] and [UMDOScenario][gemseo_umdo.scenarios.umdo_scenario.UMDOScenario]
+  when the argument ``uncertain_design_variables`` is set to noise design variables.
+  In the case of bi-level formulations,
+  it may be preferable to use this function directly rather than through this argument,
+  as illustrated in an example of the documentation.
+- The argument `n_samples_increment` of
+  the [SequentialSampling][gemseo_umdo.formulations.sampling.Sampling] U-MDO formulation
+  can be either the increment of the sampling size
+  or a function computing this increment from the current sampling size.
+- The U-MDO formulations [Surrogate][gemseo_umdo.formulations.surrogate.Surrogate]
+  and [PCERegressor][gemseo.mlearning.regression.algos.pce.PCERegressor]
+  store the quality of the surrogate models in the database attached to the scenario
+  (see `scenario.formulation.optimization_problem.database`).
+
+### Fixed
+
+- The iterative Monte Carlo estimation of statistics when used for derivatives.
+
 ## Version 4.1.0 (April 2025)
 
 ### Added